fix: check against project budget

+++ projecten3-1920-backend-klim03/Domain/Models/Domain/Order.cs
@@ -44,7 +44,7 @@ namespace projecten3_1920_backend_klim03.Domain.Models.Domain
 
         public void Approve()
         {
-            Group.PayOrder(GetOrderPrice());
+            Group.PayOrder((decimal)GetOrderPrice());
         }
 
     }
